Modularize github/mod.rs into submodules for maintainability
Extract link flow, priority configuration, and tests into separate
submodules to reduce mod.rs from 557 to 318 lines, meeting the 500-line
file size constraint.
